doc: frontpages
