+++ src/cipher/mod.rs
@@ -1,3 +1,5 @@
+//! This module contains the implementation of common ciphers
+
 mod aes;
 
 /// Block size in bits.
